+++ Templates/SharePoint-ADFS/CHANGELOG.md
@@ -7,6 +7,7 @@
 * Enable LDAPS (LDAP over SSL) on the Active Directory domain.
 * Create a new AD user to run the directory synchronization, and grant it permission "Replicate Directory Changes".
 * Create a synchronization connection in User Profile Service.
+* Remove DSC module xWebAdministration 3.3.0 and replace it with WebAdministrationDsc 4.0.0.
 
 ## Enhancements & bug-fixes - Published in August 8, 2022